refactor: Move constant out of side-effect-ful code
Allows importing it into the CLI without trying to create an SSM client.
