+++ geostore/dataset_keys.py
@@ -1,0 +1,1 @@
+DATASET_KEY_SEPARATOR = "-"
