+++ tests/stac_generators.py
@@ -4,8 +4,8 @@ from uuid import uuid4
 
 from multihash import SHA2_256
 
+from geostore.dataset_keys import DATASET_KEY_SEPARATOR
 from geostore.datasets.create import TITLE_CHARACTERS
-from geostore.datasets_model import DATASET_KEY_SEPARATOR
 from geostore.types import JsonObject
 
 from .general_generators import (
