+++ tests/test_datasets_endpoint.py
@@ -23,10 +23,10 @@ from geostore.aws_message_attributes import (
     MESSAGE_ATTRIBUTE_TYPE_ROOT,
     STRING_VALUE_KEY,
 )
+from geostore.dataset_keys import DATASET_KEY_SEPARATOR
 from geostore.datasets.create import TITLE_PATTERN
 from geostore.datasets.entrypoint import lambda_handler
 from geostore.datasets.get import get_dataset_filter, get_dataset_single, handle_get
-from geostore.datasets_model import DATASET_KEY_SEPARATOR
 from geostore.populate_catalog.task import CATALOG_FILENAME
 from geostore.resources import ResourceName
 from geostore.s3 import S3_URL_PREFIX
